# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /messages:
    get:
      summary: Get all messages
      tags:
        - messages
      responses:
        "200":
          description: A list of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      summary: Create a new message
      tags:
        - messages
      requestBody:
        description: The message to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"
      responses:
        "201":
          description: The created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}:
    patch:
      summary: Update message by id
      operationId: PatchMessageById
      tags:
        - messages
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint
          description: Id of message to update
      requestBody:
        description: Data to update message
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"
      responses:
        "200":
          description: The updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      summary: Delete message by id
      operationId: DeleteMessageById
      tags:
        - messages
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint
          description: Id of message to delete
      responses:
        "204":
          description: Message deleted successfully
  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: The created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      summary: Update user by id
      operationId: PatchUserById
      tags:
        - users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint
          description: Id of user to update
      requestBody:
        description: Data to update user
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: The updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      summary: Delete user by id
      operationId: DeleteUserById
      tags:
        - users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint
          description: Id of user to delete
      responses:
        "204":
          description: User deleted successfully

components:
  schemas:
    Message:
      type: object
      properties:
        id:
          type: integer
          format: uint
        task:
          type: string
        is_done:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
        password:
          type: string